refactor: inline detail page metadata into page.tsx
